py_analytics/requirements.txt
joblib==1.5.3
numpy==1.26.4
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2026.1.post1
pyzmq==27.1.0
river==0.21.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
